Fixes unconfigured pill presses manufacturing pills that claim to be bottles (#92832)
## About The Pull Request

Packaging category was unset until you swapped the product type in the
UI, so it'd default to naming the pills bottles. Also decapitalized the
suffixes for parity with the rest of pills/patches/bottles.

## Changelog
:cl:
fix: Fixed unconfigured pill presses manufacturing pills that claim to
be bottles
spellcheck: Decapitalized pill press suffixes for parity with the rest
of pills/patches/bottles
/:cl: