Fix force say and close the say popup instead of just clearing it (#92883)
## About The Pull Request

Changes `handleForceSay` to access the popups value via the `innerRef`
instead of the `value` variable and changes `currentPrefix` into a Ref
for similar reasons.

I can't explain very well WHY this fixes it, as I'm no typescript
developer, but I assume it has to do with how `handleForceSay` is being
used as a subscription callback. So, the other functions being used that
way may have similar issues but I'd rather not touch anymore than this.

The condition within was also incorrect but just never got anywhere
because `value` would always be empty.

This also changes it to close the say popup rather than just clearing
it. IMO it makes more sense to do this as currently this really only
happens when you are taking damage or getting stunned, so it only going
half way even though it ejected your text feels awkward.
## Why It's Good For The Game

Fixes a feature that has been seemingly broken and forgotten about for
over half a year. Maybe there's some context for why it was never fixed
idk. But, yknow, its cool and I'd like to see it in game.

![dreamseeker_fPcVtN8LkH](https://github.com/user-attachments/assets/f66461f1-ae48-4edd-b5f6-6a81de227647)
## Changelog
:cl: sushi
fix: force say works again
/:cl: