Changes the tooltip for holders of the lockable_storage component to be accurate (#92737)
## About The Pull Request

Previously, it would always say [left-click icon] open storage if you
had an empty hand. This makes it so that if the item is on the ground,
it'll say [right-click icon] open storage. if it's in your other hand,
it'll still say [left-click icon].

## Why It's Good For The Game

fixes #87340

## Changelog
:cl:

fix: tooltip for secure briefcases now shows correct button.

/:cl:

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>