Another fix for ore silo logging interface (#92610)
## About The Pull Request
This adds null-checking in the Typescript itself to create values if
there are some missing, specifically with values that indicate that this
is replacing a value that was null as a result of an error. It also
updates another place where I forgot to correctly set user_data.

## Why It's Good For The Game
Fixes more bugs.

## Changelog
:cl: Bisar
fix: The .tsx code for Ore Silo logs will now catch nulls instead of
TBing.
fix: Fixed ANOTHER place where user_data for an ore silo log wasn't set
correctly.
/:cl: