The chain of command boosts action speed and not just move speed. (#92964)
## About The Pull Request
Most contributors do not know or forget that we also have action speed
modifiers (which are a tremendous copypasta tbh) and not just move speed
modifiers. This is probably just another example of this.

Also the item had an unused variable, no context tips and didn't have an
attack animation, a balloon alert or a chat message when performing the
action, only a whip sound.

Also gave the chain of command its own file.

## Why It's Good For The Game
I think it makes sense for the chain of command to hasten someone in
ways other than just running one's legs. Also it's kinda more effective
at boosting action speed because `do_after` actions are not as frequent
as moving the mob in any direction at all.

## Changelog

:cl:
balance: The secondary action of the chain of command (the chained whip)
also boosts action speed and not just move speed.
qol: The chain of command now has contextual tip for its secondary
action, and plays the whip attack animation when it's done.
/:cl: