recipe speed fix (#92939)
## About The Pull Request
This PR aims to fix another issue with #92577. The main problem is that
the list of instances we get from `get_surroundings()` doesn't contain
items that are found within the person that's crafting the recipe, since
those are checked separately in `check_tools()`. This ends up making the
recipe hella slower because the appropriate tool couldn't be found in
the checks for calculating the duration of the craft.

This also trims a few lines of code and other small things worth of a
small refactor.

EDIT: Found out the real problem. the list of instances was an assoc
list of type -> instances, while I thought otherwise.

## Why It's Good For The Game
Fixing some recipes taking too long to make. Fixes
https://github.com/tgstation/tgstation/issues/93003

## Changelog

:cl:
fix: Fixed some recipes taking too long to make.
/:cl: