airlocks affected by dimensional anomalies retain their access (#93148)
## About The Pull Request
I'm not sure if to classify this as a fix or small balance detail, but I
don't think dimensional anomalies should nuke airlock access and other
things, such as unrestricted sides, locks and emergency access, when
transformed it into another airlock.

## Why It's Good For The Game
Dimensional anomalies are supposed to fuck up with the aesthetics of the
room. This is more of a side effect of the fact "transformed" objects
are actually deleted and new ones placed in their stead. This also
extends to dimensional bomb cores. Dimensional reactive armors don't
fuck with doors and objects, only spawn some cades and change turfs.