Adds loganuk as code owner for Catwalk map only (#93195)
## About The Pull Request

Spent a lot of time driving change for Catwalk as it has so much
potential. I'd like to be added as a reviewer mainly so I can get a ping
if there are tweaks that affect the map file during this project as well
as being able to have the chance to suggest "oooo what about this"

I have added myself to the entire folder as eventually, The map file
will be renamed from 2023 to something else but that's a future thing.

If anyone is curious about the history about feedback attempts:
https://discord.com/channels/326822144233439242/1411339428949131314

**Current and closed PRs specifically for Catwalk:**

- https://github.com/tgstation/tgstation/pull/92802
- https://github.com/tgstation/tgstation/pull/92830
- https://github.com/tgstation-operations/server-config/pull/103
- https://github.com/tgstation/tgstation/pull/93057
- https://github.com/tgstation/tgstation/pull/93058
- https://github.com/tgstation/tgstation/pull/93150
## Why It's Good For The Game

- I can do it without this, this will just make it a bit easier for me
to keep on top of it as merge conflicts are a plenty at the moment