Lobby TV screen shows an icon of the overflow job (#93023)
## About The Pull Request

Hey so remember that TV thing I added, and I wanted it to be useful for
overflow jobs so I added it in, but couldn't make it look good with
runechat so I made it only appear post-roundstart, rendering it utterly
useless?

Yeah, I do too, so I've come back to fix it with the help of stealing
other people's code (I'm a lazy individual) - Taking Absolucy's work in
https://github.com/Monkestation/Monkestation2.0/pull/4589 for this.

Now it shows an icon of the job with an active overflow under the TV's
effects, and has a tooltip if you hover over in case you don't
understand what it's demonstrating.

https://github.com/user-attachments/assets/97489f93-f40d-47f9-aea2-6cc32c8fcba3

## Why It's Good For The Game

The TV shows overflow information before the round starts, allowing
people to actually see and react to the information rather than only
being shown to latejoiners. I prefer it this way as I don't want to have
to ensure the different things we may put on the TV's runetext all shows
up properly.

## Changelog

:cl: Absolucy, JohnFulpWillard
image: The Lobby TV now shows an icon of the overflow job.
/:cl: