Fixes harddels and UI-breaking runtimes from open_containers (#93230)
## About The Pull Request

Closes #93221 as it incorporates a better version of the fix in said PR.
hud_used could exist without a client (such as if the player logged out
temporarily) and the elements in the list were not removed when an item
was taken out of storage. This could result in harddels, which would
fill the list with nulls and lead to the following runtime which would
prevent all further UI updates on the mob, including planemaster
rebuilds

```js
The following runtime has occurred 161 time(s).
runtime error: wrong type of value for list
proc name: show hud (/datum/hud/proc/show_hud)
  source file: code/_onclick/hud/hud.dm,347
  usr: Doctors-The-Smiles (/mob/living/carbon/human)
  src: /datum/hud/human (/datum/hud/human)
```

Fixes #93212

~~This is ***not*** a solution to #93212 (or at least I think so, there
is a separate filter issue), but it could also break planemasters on a
mob.~~

## Changelog
:cl:
fix: Fixed harddels and UI-breaking runtimes from open_containers
/:cl: