Fixes clicking on an open airlock tile when pulling something (#93317)
## About The Pull Request

If you pull an item, and then click on an open airlock tile, it will
close the door instead of dragging the pulled item onto the tile. This
PR fixes that.

You can still click on the door with LMB.

The change I've made, shown by only left clicking demonstration:

https://github.com/user-attachments/assets/11f294b1-5cdd-458a-8757-fd3edd279ffd

^*It used to close the airlock when clicking on the empty tile. It works
as intended in the video demonstration.

## Why It's Good For The Game

When dragging a person, I sometimes have wanted to sometimes drag
objects/people onto the airlock tile, but instead, it would close the
airlock. I found this frustrating because my intention wasn't to close
the airlock.

With this PR, I've made it so when pulling an object/person it will not
close the airlock. it will instead do what was intended, and moves the
pulled object toward the clicked airlock tile.

You can still close doors through LMB on the door, so the original
functionality is still easily available.

## Changelog

:cl: ArchBTW
fix: fixes clicking on an open airlock tile when pulling something
/:cl:

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>