Removed duplicate scream emote. (#93104)
## About The Pull Request
We have two different *scream emote types, not counting the *screech
subtype, and it's a bit of a copypasta, so I merged the two together,
and made the *screech its own type, only usable by human mobs ~~with the
primitive traits, aka monkeys and ashwalkers~~ like before.

## Why It's Good For The Game
Less copypasta. You'll probably have to reset your scream hotkey after
this is merged.

~~Also screeching is simply another emote that played the scream sound,
and human scream stock sounds barely fit it anyway, so we're not losing
much by locking it to monkeys and ashwalkers.~~ Screeching is once more
back on the menu

## Changelog

:cl:
code: Merged the human and living versions of the scream emote. Remember
to reset your scream hotkey if it is ever unset.
/:cl: