syndicate modsuit visor pseudo-toggle (tied to combat mode) (#93391)
## About The Pull Request

The visors on MODsuits using the Syndicate welding visor (now renamed
the "flash-protected optical suite") now toggle their armor booster
overlay depending on whether the user has combat mode on or off.

## Why It's Good For The Game

Style. Listen, man, I just thought it'd be a shame if the cool visor
sprites never got used again.

https://github.com/user-attachments/assets/d5103168-7a1e-4405-adb3-53420c8601d7

## Changelog

:cl:
add: The visors on MODsuits that used to have armor boosters now visibly
toggle based on whether the wearer is in combat mode or not.
/:cl:

---------

Co-authored-by: Hatterhat <Hatterhat@users.noreply.github.com>
Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>