The Inversenning : Chiral Inversing Buffer (#93376)
## About The Pull Request

This PR adds a new buffer (and accompanying Lavaland geyser) to the
game: Chiral Inversing Buffer! This unique buffer allows you to
instantly transmute a hidden impure or inverse reagent into its original
form, no hassle with the HPLC. It can be acquired in three ways, two of
which involve Cargo. You can purchase a crate containing one 30u bottle
of the buffer for 600 credits, or for free through Medbay's cargo
console. You can also scour Lavaland and uncover a geyser that offers
unlimited buffer.

The buffer itself checks if any eligible inverses or impurities are
present within the holder. If there is none, the buffer will fizzle out
and be effectively wasted. If there is an eligible reagent, it converts
the reagent into the Inverse at a 10-to-1 ratio with the buffer. This
means that buying a bottle from cargo will net you at max, 300 units of
instant inversing. Do note that this buffer has the best results when
added after a reaction, not during or before.

https://github.com/user-attachments/assets/5f04097d-6baa-4cfd-bf0d-77f175d37acb

## Why It's Good For The Game

This was made from the suggestions of my previous PR at #93316. The main
issue that this PR targets is the lack of speedy access to "true"
inverse reagents. The current system hides the name of most inverses on
creation and only has the HPLC to reliably convert hidden inverses into
their true counterpart. This can cause issues with some inverses due to
them still retaining some properties of their base reagent. While that
system cannot be removed as of now, this PR serves to add another
alternative for Chemists to pursue if they need to purify their inverse
outputs.

<img width="135" height="136" alt="chiralgeyser"
src="https://github.com/user-attachments/assets/1cb336ac-966e-4c4b-8cb2-68616c410f26"
/>

## Changelog
:cl:
add: Chiral Inversing Buffer : A new way to convert your hidden
impurities and inverses into their true form! Requires a purchase from
Cargo or a trip to Lavaland to acquire.
balance: A new geyser has been added to Lavaland that outputs Chiral
Inversing Buffer. Geyser weights have been adjusted to accommodate this
addition.
/:cl: